+++ R/tar_engine_knitr.R
@@ -1,14 +1,38 @@
 #' @title Target Markdown `knitr` engine
 #' @export
-#' @family engines
+#' @family Target Markdown
 #' @seealso <https://books.ropensci.org/targets/markdown.html>
 #' @description `knitr` language engine that runs `{targets}`
 #'   code chunks in Target Markdown.
 #' @return Character, output generated from `knitr::engine_output()`.
 #' @param options A named list of `knitr` chunk options.
+#' @section Target Markdown interactive mode:
+#'   Target Markdown has two modes:
+#'   1. Non-interactive mode. This is the default when you
+#'     run `knitr::knit()` or `rmarkdown::render()`.
+#'     Here, the code in `{targets}` code chunks gets written
+#'     to special script files in order to set up a `targets`
+#'     pipeline to run later.
+#'   2. Interactive mode: here, no scripts are written to set up
+#'     a pipeline. Rather, the globals or targets in question
+#'     are run in the current environment and the values
+#'     are assigned to that environment.
+#'
+#'   The mode is interactive if `!isTRUE(getOption("knitr.in.progress"))`,
+#'   is `TRUE`. The `knitr.in.progress` option is `TRUE`
+#'   when you run `knitr::knit()` or `rmarkdown::render()`
+#'   and `NULL` if you are running one chunk at a time interactively
+#'   in an integrated development environment, e.g. the
+#'   notebook interface in RStudio:
+#'   <https://bookdown.org/yihui/rmarkdown/notebook.html>.
+#'   You can choose the mode with the `tar_interactive`
+#'   chunk option.
+#'   (In `targets` 0.6.0, `tar_interactive` defaults to `interactive()`
+#'   instead of `!isTRUE(getOption("knitr.in.progress"))`.)
 #' @section Target Markdown chunk options:
 #'   Target Markdown introduces the following `knitr` code chunk options.
-#'   Most other standard `knitr` code chunk options should just work.
+#'   Most other standard `knitr` code chunk options should just work
+#'   in non-interactive mode. In interactive mode, not all
 #'   * `tar_globals`: Logical of length 1,
 #'     whether to define globals or targets.
 #'     If `TRUE`, the chunk code defines functions, objects, and options
@@ -16,7 +40,8 @@
 #'     then the chunk returns formal targets for the pipeline.
 #'   * `tar_interactive`: Logical of length 1, whether to run in
 #'     interactive mode or non-interactive mode.
-#'     Defaults to the return value of `interactive()`.
+#'     See the "Target Markdown interactive mode" section of this
+#'     help file for details.
 #'   * `tar_name`: name to use for writing helper script files
 #'     (e.g. `_targets_r/targets/target_script.R`)
 #'     and specifying target names if the `tar_simple` chunk option
@@ -62,29 +87,25 @@ tar_engine_knitr <- function(options) {
   if (identical(as.logical(options$eval), FALSE)) {
     return(engine_knitr_output(options = options, out = character(0)))
   }
-  tar_assert_package("knitr")
-  tar_assert_list(options, "knitr chunk options must be a list.")
   options$tar_name <- options$tar_name %|||% options$label
-  msg <- paste(
-    "{targets} code chunks require a nonempty length-1 character string",
-    "for the chunk label or the tar_name chunk option."
-  )
-  tar_assert_scalar(options$tar_name, msg)
-  tar_assert_chr(options$tar_name, msg)
-  tar_assert_nzchar(options$tar_name, msg)
+  options$tar_script <- options$tar_script %|||% tar_config_get("script")
   if (!is.null(options$targets)) {
+    options$tar_globals <- options$tar_globals %|||% options$targets
     tar_warn_deprecate(
-      "In Target Markdown, the `targets` chunk option is deprecated.",
+      "In Target Markdown, the `targets` chunk option is deprecated ",
+      "(version 0.6.0, 2021-07-21). ",
       "Set the chunk option tar_globals = TRUE to define functions, ",
       "global objects, and settings. To define targets, ",
       "either set tar_globals = FALSE or leave tar_globals unset."
     )
-    options$tar_globals <- options$tar_globals %|||% options$targets
   }
-  options$tar_script <- options$tar_script %|||% tar_config_get("script")
-  engine_knitr_tar_assert_options(options)
+  tar_assert_package("knitr")
+  tar_assert_list(options, "knitr chunk options must be a list.")
   warn_labels_duplicated()
   warn_labels_unnamed(options)
+  engine_knitr_tar_assert_options(options)
+  engine_knitr_set_interactive(options)
+  on.exit(engine_knitr_unset_interactive())
   if_any(
     identical(options$tar_globals, TRUE),
     engine_knitr_globals(options),
@@ -93,30 +114,59 @@ tar_engine_knitr <- function(options) {
 }
 
 engine_knitr_tar_assert_options <- function(options) {
-  choices <- c("tar_globals", "tar_interactive", "tar_script", "tar_simple")
-  for (option in choices) {
+  scalars <- c(
+    "tar_globals",
+    "tar_interactive",
+    "tar_name",
+    "tar_script",
+    "tar_simple"
+  )
+  logicals <- c("tar_globals", "tar_interactive", "tar_simple")
+  characters <- c("tar_name", "tar_script")
+  for (option in scalars) {
     tar_assert_scalar(
       options[[option]] %|||% TRUE,
       paste(option, "chunk option must either be NULL or have length 1.")
     )
   }
-  for (option in c("tar_globals", "tar_interactive", "tar_simple")) {
+  for (option in logicals) {
     tar_assert_lgl(
       options[[option]] %|||% TRUE,
       paste(option, "chunk option must either be NULL or logical.")
     )
   }
-  tar_assert_chr(
-    options[["tar_script"]],
-    "tar_script chunk option must either be NULL or character."
-  )
+  for (option in characters) {
+    tar_assert_chr(
+      options[[option]],
+      paste(option, "chunk option must either be NULL or character.")
+    )
+    tar_assert_nzchar(
+      options[[option]],
+      paste(option, "chunk option must not be the empty string.")
+    )
+  }
+}
+
+engine_knitr_set_interactive <- function(options) {
+  x <- options$tar_interactive %|||% !isTRUE(getOption("knitr.in.progress"))
+  tar_assert_scalar(x)
+  tar_assert_lgl(x)
+  tar_runtime$set_interactive(x)
+}
+
+engine_knitr_unset_interactive <- function() {
+  tar_runtime$unset_interactive()
+}
+
+engine_knitr_is_interactive <- function() {
+  isTRUE(tar_runtime$get_interactive())
 }
 
 engine_knitr_globals <- function(options) {
   if_any(
-    options$tar_interactive %|||% interactive(),
-    engine_knitr_globals_prototype(options),
-    engine_knitr_globals_construct(options)
+    engine_knitr_is_interactive(),
+    engine_knitr_globals_interactive(options),
+    engine_knitr_globals_noninteractive(options)
   )
 }
 
@@ -125,9 +175,9 @@ engine_knitr_targets <- function(options) {
     options$code <- engine_knitr_targets_command(options)
   }
   if_any(
-    options$tar_interactive %|||% interactive(),
-    engine_knitr_targets_prototype(options),
-    engine_knitr_targets_construct(options)
+    engine_knitr_is_interactive(),
+    engine_knitr_targets_interactive(options),
+    engine_knitr_targets_noninteractive(options)
   )
 }
 
@@ -139,19 +189,27 @@ engine_knitr_targets_command <- function(options) {
   )
 }
 
-engine_knitr_globals_prototype <- function(options) {
-  eval(parse(text = options$code), envir = tar_option_get("envir"))
-  engine_knitr_output(
-    options,
-    "Ran code and assigned objects to the environment."
+engine_knitr_globals_interactive <- function(options) {
+  out_code <- engine_knitr_echo_code(options)
+  message <- "Run code and assign objects to the environment."
+  out_message <- engine_knitr_run_message(options, message)
+  options_globals <- options
+  options_globals$echo <- FALSE
+  code <- c(
+    "evalq({",
+    options_globals$code,
+    "}, envir = targets::tar_option_get(\"envir\"))"
   )
+  options_globals$code <- paste(code, collapse = "\n")
+  out_globals <- knitr::knit_engines$get("R")(options = options_globals)
+  paste0(out_code, out_message, out_globals)
 }
 
-engine_knitr_globals_construct <- function(options) {
+engine_knitr_globals_noninteractive <- function(options) {
   write_targets_r(options$tar_script)
   write_targets_r_globals(options$code, options$tar_name, options$tar_script)
   out <- paste0(
-    "Established ",
+    "Establish ",
     options$tar_script,
     " and ",
     path_script_r_globals(options$tar_script, options$tar_name),
@@ -160,20 +218,30 @@ engine_knitr_globals_construct <- function(options) {
   engine_knitr_output(options, out)
 }
 
-engine_knitr_targets_prototype <- function(options) {
-  tar_make_interactive(options$code)
-  out <- c(
+engine_knitr_targets_interactive <- function(options) {
+  out_code <- engine_knitr_echo_code(options)
+  message <- paste(
     engine_knitr_definition_message(options),
-    engine_knitr_prototype_message(options)
+    engine_knitr_interactive_message(options),
+    sep = "\n"
   )
-  engine_knitr_output(options, out)
+  message <- trimws(message)
+  out_message <- engine_knitr_run_message(options, message)
+  options_make <- options
+  code_library <- "library(targets)"
+  code_make <- c("targets::tar_make_interactive(", deparse(options$code), ")")
+  code_make <- paste(code_make, collapse = "")
+  options_make$code <- paste(code_library, code_make, sep = "\n")
+  options_make$echo <- FALSE
+  out_make <- knitr::knit_engines$get("R")(options = options_make)
+  paste0(out_code, out_message, out_make)
 }
 
-engine_knitr_targets_construct <- function(options) {
+engine_knitr_targets_noninteractive <- function(options) {
   write_targets_r(options$tar_script)
   write_targets_r_targets(options$code, options$tar_name, options$tar_script)
   out <- paste0(
-    "Established ",
+    "Establish ",
     options$tar_script,
     " and ",
     path_script_r_targets(options$tar_script, options$tar_name),
@@ -183,29 +251,42 @@ engine_knitr_targets_construct <- function(options) {
   engine_knitr_output(options, out)
 }
 
+engine_knitr_echo_code <- function(options) {
+  options$eval <- FALSE
+  options$results <- "hide"
+  knitr::knit_engines$get("R")(options)
+}
+
+engine_knitr_run_message <- function(options, message) {
+  options$code <- sprintf("message(\"%s\")", message)
+  options$echo <- FALSE
+  knitr::knit_engines$get("R")(options)
+}
+
 engine_knitr_output <- function(options, out) {
   code <- paste(options$code, collapse = "\n")
   options$engine <- "r"
+  out <- if_any(options$message, out, character(0))
   knitr::engine_output(options = options, code = code, out = out)
 }
 
 engine_knitr_definition_message <- function(options) {
   if_any(
     options$tar_simple %|||% FALSE,
-    paste("Defined target", options$tar_name, "automatically from chunk code."),
+    paste("Define target", options$tar_name, "from chunk code."),
     character(0)
   )
 }
 
-engine_knitr_prototype_message <- function(options) {
+engine_knitr_interactive_message <- function(options) {
   if_any(
     options$tar_simple %|||% FALSE,
     paste(
-      "Ran target",
+      "Run target",
       options$tar_name,
-      "and assigned it to the environment."
+      "and assign it to the environment."
     ),
-    "Ran targets and assigned them to the environment."
+    "Run targets and assign them to the environment."
   )
 }
 
@@ -271,8 +352,9 @@ warn_labels_duplicated <- function() {
       "interfere with the proper execution of Target Markdown ",
       "unless you set unique values for the tar_name chunk option. ",
       "Please set knitr.duplicate.label to a value other than \"allow\" ",
-      "to prohibit duplicate knitr chunk labels. Suppress this warning with ",
-      "Sys.setenv(TAR_WARN = \"false\")."
+      "to prohibit duplicate knitr chunk labels. ",
+      "Warnings like this one are important, but if you must suppress them, ",
+      "you can do so with Sys.setenv(TAR_WARN = \"false\")."
     )
   }
 }
@@ -283,7 +365,8 @@ warn_labels_unnamed <- function(options) {
     tar_warn_validate(
       "Please assign explicit labels to {targets} code chunks ",
       "in order to avoid accidental duplicated script files. ",
-      "Suppress this warning with Sys.setenv(TAR_WARN = \"false\")."
+      "Warnings like this one are important, ",
+      "but you can suppress them with Sys.setenv(TAR_WARN = \"false\")."
     )
   }
 }