+++ R/tar_watch.R
@@ -66,7 +66,10 @@ tar_watch <- function(
   host = getOption("shiny.host", "127.0.0.1"),
   port = getOption("shiny.port", targets::tar_random_port()),
   verbose = TRUE,
-  supervise = TRUE
+  supervise = TRUE,
+  poll_connection = TRUE,
+  stdout = "|",
+  stderr = "|"
 ) {
   tar_assert_watch_packages()
   tar_assert_chr(exclude)
@@ -117,9 +120,10 @@ tar_watch <- function(
   process <- callr::r_bg(
     func = tar_watch_app,
     args = args,
-    stdout = "|",
-    stderr = "|",
-    supervise = supervise
+    stdout = stdout,
+    stderr = stderr,
+    supervise = supervise,
+    poll_connection = poll_connection
   )
   if (browse) {
     url_port(host = host, port = port,  process = process, verbose = verbose)